general.js a utils.mjs en pagina de perfil
