+++ src/public/perfil.js
@@ -69,6 +69,10 @@ import utils from '/modules/utils.mjs';
     levelMessageNavbar.childNodes[2].nodeValue = `¡Felicidades! Has alcanzado el nivel ${gameData.level}`;
     progressBarColor.style.width = `${gameData.experience}%`;
 
+    utils.showAdminNavbar();
     utils.showPageElements();
+    utils.showNotifications();
+    utils.navbarUserName();
+    utils.logOutButtons();
 
 })();
